Centralize linksList
