+++ src/front/component/visualcomponents/Hamburger.js
@@ -5,16 +5,8 @@ import { NavBarLinksList } from './NavBarLinksList'
 import { useTranslation } from 'react-i18next'
 import { Footer } from '../../pages/Footer'
 
-export const Hamburger = () => {
-
+export const Hamburger = ({ linksList }) => {
   const { t } = useTranslation("global");
-
-  const linksList = {
-    "introduction-section": t("navbar.aboutus"),
-    "unlock-section": t("navbar.services"),
-    "team-section": t("navbar.team"),
-    "contact-section": t("navbar.contact")
-  };
 
   return (<>
     <label htmlFor="menu-control" className="hamburger">
