+++ src/front/component/visualcomponents/NavBar.js
@@ -5,17 +5,8 @@ import { ContactButton } from './ContactButton'
 import { LanguageDropdown } from './LanguageDropdown'
 import { useTranslation } from 'react-i18next'
 
-
-export const NavBar = () => {
+export const NavBar = ({ linksList }) => {
   const { t } = useTranslation("global");
-
-  const linksList = {
-    "introduction-section": t("navbar.aboutus"),
-    "unlock-section": t("navbar.services"),
-    "team-section": t("navbar.team"),
-    "contact-section": t("navbar.contact"),
-    "blog-page" : "Blog",
-  };
 
   return (
     <nav className='navbar-container'>
@@ -24,7 +15,12 @@ export const NavBar = () => {
       </div>
       <div className='navbar-buttons-section'>
         <LanguageDropdown />
-        <ContactButton buttonClass='landing-introduction-booking-button' buttonText={t("buttons.appointment")} buttonType="CALENDLY" showButton={true} />
+        <ContactButton
+          buttonClass='landing-introduction-booking-button'
+          buttonText={t("buttons.appointment")}
+          buttonType="CALENDLY"
+          showButton={true}
+        />
       </div>
     </nav>
   )
